refactor(v0): extract session-state query service
